requirements.txt
flask==2.0.1
docxtpl==0.11.5
Werkzeug==2.0.1
python-docx==0.8.11
requests==2.26.0
azure-identity==1.7.0
azure-keyvault-secrets==4.3.0